requirements.txt
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
Pillow==8.3.1
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
queuelib==1.6.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
Scrapy==2.5.0
scrapy-redis==0.7.1
scrapy-selenium==0.0.7
selenium==3.141.0
Selenium-Screenshot==1.7.0
service-identity==21.1.0
six==1.16.0
Twisted==21.2.0
urllib3==1.26.6
w3lib==1.22.0
zope.interface==5.4.0
